requirements.txt
# Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-decouple==3.8
psycopg2-binary==2.9.9  # ou use sqlite para desenvolvimento
drf-spectacular==0.27.1  # opcional, para documentação OpenAPI